requirements.txt
Flask==2.2.2
gunicorn==20.1.0
python-dotenv==0.15.0